Moved queue creation to util
